ai_personality_project/requirements.txt
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
scikit-learn>=1.2.0
gradio>=4.21.0
sentence-transformers>=2.2.0
flask>=2.3.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.10.0
pandas>=2.0.0
matplotlib>=3.7.0